# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API",
        "version": "1.0"
    },
    "paths": {
...
        "/dashboard/first_login": {
            "post": {
                "tags": [
                    "Livret"
                ],
                "summary": "Crée un nouveau livret pour l'utilisateur authentifié.",
                "description": "Crée un nouveau livret pour l'utilisateur authentifié.",
                "operationId": "store",
                "requestBody": {
                    "description": "Données du livret à créer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "livret_name",
                                    "establishment_type",
                                    "establishment_name",
                                    "establishment_address",
                                    "establishment_phone",
                                    "establishment_email",
                                    "establishment_website"
                                ],
                                "properties": {
                                    "livret_name": {
                                        "type": "string",
                                        "example": "Mon livret"
                                    },
                                    "establishment_type": {
                                        "type": "string",
                                        "example": "Hôtel"
                                    },
                                    "establishment_name": {
                                        "type": "string",
                                        "example": "Hôtel de la plage"
                                    },
                                    "establishment_address": {
                                        "type": "string",
                                        "example": "12 rue de la plage"
                                    },
                                    "establishment_phone": {
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "establishment_email": {
                                        "type": "string",
                                        "example": "email@email.com"
                                    },
                                    "establishment_website": {
                                        "type": "string",
                                        "example": "https://www.hotelplage.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livret créé avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Livret créé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation des données.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "livret_name": [
                                                    "Le champ nom du livret est obligatoire."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/livret/{slug}/{id}": {
            "get": {
                "tags": [
                    "Livret"
                ],
                "summary": "Affiche un livret spécifique selon son slug et son ID.",
                "description": "Affiche un livret spécifique selon son slug et son ID.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Le slug du livret.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'ID du livret.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livret trouvé.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "livret": {
                                            "type": "object"
                                        },
                                        "modules": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livret introuvable.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ce livret n'existe pas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}